Allow associating items with arbitrary data
This commit adds a feature to associate cart items with arbitrary data.
The feature may be useful to store information that isn't available
through the item's model instance, such as the date and time the item
was added to the cart.

It's implemented by allowing the item class constructor to accept extra
keyword arguments, which are used to set new attributes on the item.
These arguments are meant to be passed through the cart's method add().

Examples
--------
>>> item = BaseItem(obj, foo='bar')
>>> item.foo
'bar

>>> cart.add(pk=1, timestamp=time.time())
>>> item = cart.items['1']
>>> item.timestamp
1463051568.704125